+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 # PiperLang Changelog
 
 ## WIP - Next Version
-* Nothing to see yet
+* Nothing to see yet!
+
+## Version 1.0 Beta 2 (08-04-24)
+* Added cookie enabled setting (closes #14)
+* Added show symbol option to currencyFormat method
+* Added setLanguageSession method
+* Added setLanguageCookie method
+* No longer set the language inside the constructor
+* Improved detectUserLanguage method
+* Improved format handling inside the dateFormat method
+* Improved format error messages
+* Updated docs in README.md
+* Fixed some style and typo things
 
 ## Version 1.0 Beta 1 (04-04-24)
 * Initial release